pyproject.toml
[project]
name = "nautilus-trial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nautilus-trader==1.224.0",
    "py-clob-client>=0.19.0",
    "python-dotenv>=1.0.0",
    "web3>=6.0.0",
]

[tool.setuptools.packages.find]
exclude = ["live*", "historical*", "data*", "docs*"]
